fix: restrict scheduling to the future
- this wasn't working previously
- tidy up use of datetime-local across cms views
