+++ .changeset/six-wasps-divide.md
@@ -1,0 +1,8 @@
+---
+'@blinkk/root-cms': patch
+---
+
+fix: restrict scheduling to the future
+
+- this wasn't working previously
+- tidy up use of datetime-local across cms views
